+++ package.json
@@ -8,12 +8,13 @@
     "dev": "nuxt",
     "build": "nuxt build",
     "start": "nuxt start",
-    "generate": "node -r esm build-index.js; nuxt generate",
+    "generate": "node -r esm build-data.js; node -r esm build-index.js; nuxt generate",
     "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
     "precommit": "npm run lint",
     "test": "jest",
     "deploy": "push-dir --dir=dist --branch=gh-pages --cleanup",
-    "index": "node -r esm build-index.js"
+    "index": "node -r esm build-index.js",
+    "data": "node -r esm build-data.js"
   },
   "dependencies": {
     "@nuxtjs/google-analytics": "^2.0.2",
