+++ static/data/changelog.json
@@ -1,10 +1,10 @@
 [
   {
     "layout": "post",
-    "title": "Version 1.1.1",
-    "description": "Want to search they player's manual? Now you can!",
+    "title": "Site Search is here!",
+    "description": "Want to search they player's manual? Now you can...",
     "date": "2019-05-13T00:00:00.000Z",
-    "image": "https://vignette.wikia.nocookie.net/masseffect/images/8/86/Planetscanning.png/revision/latest/scale-to-width-down/350?cb=20100201034738",
+    "image": "https://vignette.wikia.nocookie.net/masseffect/images/8/86/Planetscanning.png/revision/latest?cb=20100201034738",
     "slug": "1.1.1",
     "url": "/changelog/1.1.1"
   },
